LayoutTests/http/tests/security/import-module-crossorigin-loads-expected.txt

ALERT: script ran.
This test passes if the module script loads.

PASS
